Organize UI elements into new package
