+++ src/main/java/com/flippingutilities/FlippingItem.java
@@ -26,6 +26,7 @@
 
 package com.flippingutilities;
 
+import com.flippingutilities.ui.FlippingItemPanel;
 import java.time.Instant;
 import lombok.Getter;
 import lombok.Setter;
