+++ src/main/java/com/flippingutilities/ui/FlippingPanel.java
@@ -24,8 +24,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.flippingutilities;
+package com.flippingutilities.ui;
 
+import com.flippingutilities.FlippingItem;
+import com.flippingutilities.FlippingPlugin;
 import com.google.common.base.Strings;
 import java.awt.BorderLayout;
 import java.awt.CardLayout;
