+++ src/main/java/com/flippingutilities/ui/TabManager.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.flippingutilities;
+package com.flippingutilities.ui;
 
 import java.awt.BorderLayout;
 import javax.inject.Inject;
